Rename upgrade files to update
