+++ test/e2e/contracts/updatable-contract-v2/contract/tsconfig.json
