+++ test/e2e/contracts/updatable-contract/contract/UpdatableContract.ts
@@ -1,4 +1,5 @@
 import {
+    Address,
     Blockchain,
     BytesWriter,
     Calldata,
@@ -6,7 +7,7 @@ import {
 } from '@btc-vision/btc-runtime/runtime';
 
 @final
-export class UpgradeableContractV2 extends OP_NET {
+export class UpdatableContract extends OP_NET {
     public constructor() {
         super();
     }
@@ -15,8 +16,15 @@ export class UpgradeableContractV2 extends OP_NET {
     @returns('uint32')
     public getValue(_: Calldata): BytesWriter {
         const result = new BytesWriter(4);
-        result.writeU32(2);
+        result.writeU32(1);
         return result;
+    }
+
+    @method('address')
+    public upgrade(calldata: Calldata): BytesWriter {
+        const sourceAddress = calldata.readAddress();
+        Blockchain.updateContractFromExisting(sourceAddress);
+        return new BytesWriter(0);
     }
 
     @method('bytes32', 'bytes32')
