+++ test/e2e/contracts/updatable-contract/contract/asconfig.json
