+++ test/e2e/contracts/updatable-contract/contract/index.ts
@@ -1,5 +1,5 @@
 import { Blockchain } from '@btc-vision/btc-runtime/runtime';
-import { UpgradeableContract } from './UpgradeableContract';
+import { UpdatableContract } from './UpdatableContract';
 import { revertOnError } from '@btc-vision/btc-runtime/runtime/abort/abort';
 
 // DO NOT TOUCH TO THIS.
@@ -7,7 +7,7 @@ Blockchain.contract = () => {
     // ONLY CHANGE THE CONTRACT CLASS NAME.
     // DO NOT ADD CUSTOM LOGIC HERE.
 
-    return new UpgradeableContract();
+    return new UpdatableContract();
 };
 
 // VERY IMPORTANT
